+++ lib/plugins/aws/package/compile/events/eventBridge/index.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const _ = require('lodash');
-const crypto = require('crypto');
 const { addCustomResourceToService } = require('../../../../customResources');
 const ServerlessError = require('../../../../../../serverless-error');
+const { isCFFunc, makeAndHashRuleName, makeEventBusTargetId, makeRuleName } = require('./helpers');
 
 class AwsCompileEventBridgeEvents {
   constructor(serverless, options) {
@@ -249,10 +249,6 @@ class AwsCompileEventBridgeEvents {
                 };
 
                 _.merge(compiledCloudFormationTemplate.Resources, {
-                  // TODO: This should be the logicalId for a lambdaPermission - cannot use eventBusName here as it may CF func object
-                  // TODO: This is failing tests - need to sort the name out!!
-                  // Lambda permission unique to the lambda and the eventBus <- logical id must reflect this
-                  // Use name if string or extract logical Id if CF func?
                   [this.provider.naming.getEventBridgeLambdaPermissionLogicalId(
                     functionName,
                     idx
@@ -294,112 +290,57 @@ class AwsCompileEventBridgeEvents {
 
     // These permissions are for the custom resource lambda
     // The use should ensure the IAM role the serverless framework use has these when using native CloudFormation
-    if (useCustomResources) {
-      if (anyFuncUsesEventBridge) {
-        const ruleResources = {
-          'Fn::Join': [
-            ':',
-            [
-              'arn',
-              { Ref: 'AWS::Partition' },
-              'events',
-              { Ref: 'AWS::Region' },
-              { Ref: 'AWS::AccountId' },
-              'rule/*',
-            ],
+    if (useCustomResources && anyFuncUsesEventBridge) {
+      const ruleResources = {
+        'Fn::Join': [
+          ':',
+          [
+            'arn',
+            { Ref: 'AWS::Partition' },
+            'events',
+            { Ref: 'AWS::Region' },
+            { Ref: 'AWS::AccountId' },
+            'rule/*',
           ],
-        };
-        iamRoleStatements.push({
-          Effect: 'Allow',
-          Resource: ruleResources,
-          Action: [
-            'events:PutRule',
-            'events:RemoveTargets',
-            'events:PutTargets',
-            'events:DeleteRule',
+        ],
+      };
+      iamRoleStatements.push({
+        Effect: 'Allow',
+        Resource: ruleResources,
+        Action: [
+          'events:PutRule',
+          'events:RemoveTargets',
+          'events:PutTargets',
+          'events:DeleteRule',
+        ],
+      });
+      const functionResources = {
+        'Fn::Join': [
+          ':',
+          [
+            'arn',
+            { Ref: 'AWS::Partition' },
+            'lambda',
+            { Ref: 'AWS::Region' },
+            { Ref: 'AWS::AccountId' },
+            'function',
+            '*',
           ],
-        });
-        const functionResources = {
-          'Fn::Join': [
-            ':',
-            [
-              'arn',
-              { Ref: 'AWS::Partition' },
-              'lambda',
-              { Ref: 'AWS::Region' },
-              { Ref: 'AWS::AccountId' },
-              'function',
-              '*',
-            ],
-          ],
-        };
-        iamRoleStatements.push({
-          Effect: 'Allow',
-          Resource: functionResources,
-          Action: ['lambda:AddPermission', 'lambda:RemovePermission'],
-        });
-      }
+        ],
+      };
+      iamRoleStatements.push({
+        Effect: 'Allow',
+        Resource: functionResources,
+        Action: ['lambda:AddPermission', 'lambda:RemovePermission'],
+      });
+    }
 
-      if (iamRoleStatements.length) {
-        return addCustomResourceToService(this.provider, 'eventBridge', iamRoleStatements);
-      }
+    if (iamRoleStatements.length) {
+      return addCustomResourceToService(this.provider, 'eventBridge', iamRoleStatements);
     }
 
     return null;
   }
 }
-
-const makeAndHashRuleName = ({ functionName, index }) => {
-  const name = makeRuleName({ functionName, index });
-  if (name.length > 64) {
-    // Rule names cannot be longer than 64.
-    // Temporary solution until we have https://github.com/serverless/serverless/issues/6598
-    return hashName(name, makeRuleNameSuffix(index));
-  }
-  return name;
-};
-
-const makeRuleName = ({ functionName, index }) => `${functionName}-${makeRuleNameSuffix(index)}`;
-
-const makeRuleNameSuffix = index => `rule-${index}`;
-
-const makeEventBusTargetId = ruleName => {
-  const suffix = 'target';
-  let targetId = `${ruleName}-${suffix}`;
-  if (targetId.length > 64) {
-    // Target ids cannot be longer than 64.
-    targetId = hashName(targetId, suffix);
-  }
-  return targetId;
-};
-
-const hashName = (name, suffix) =>
-  `${name.slice(0, 31 - suffix.length)}${crypto
-    .createHash('md5')
-    .update(name)
-    .digest('hex')}-${suffix}`;
-
-const isCFFunc = data => {
-  if (typeof data !== 'object') {
-    return false;
-  }
-  const keys = Object.keys(data);
-  if (keys.length !== 1) {
-    return false;
-  }
-  const cfFuncs = ['Fn::GetAtt', 'Fn::ImportValue', 'Fn::Ref'];
-  if (!cfFuncs.includes(keys[0])) {
-    return false;
-  }
-  return true;
-};
-
-// Possible paths
-// `useCF` property is not set - fork to old code
-// `useCF` is set to true - follow below pathways
-// No eventBus set in config = use default event bus -> no name set on rule
-// TODO: Don't care about arn strings? eventBus is set as arn string = manually compose arn and name when required - use custom resource
-// eventBus is set as a CF Func = validate CF Func -> extract logicalId from func and use in of CF Resources
-// eventBus is set but not an arn string or CF Func = use plain CloudFormation -> use value of eventBridge as EventBus name
 
 module.exports = AwsCompileEventBridgeEvents;
